Moving utils into config/utils
